Set default prefix value
There must be the same number of path prefixes as config keys. Unfortunately the default CLI args contain a single path prefix, but no subkey, which is invalid and results in startup failing with `level=fatal msg="keys-prefix lengths are not equal [v1-release] [] 0.0.0.0:8080"`.

Fix that by adding a default subkey of an empty string - so just the root of the document.

Signed-off-by: Brad Davidson <brad.davidson@rancher.com>